travel-agent-api/pyproject.toml
[project]
name = "travel-agent-api"
version = "0.1.0"
description = ""
authors = [
    {name = "salvatoretroiano",email = "luca-troiano@live.it"}
]
readme = "README.md"
requires-python ="<4.0,>=3.9.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "openai (>=1.85.0,<2.0.0)",
    "langchain-openai (>=0.3.21,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "serpapi (>=0.1.5,<0.2.0)",
    "pydantic (>=2.11.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "travel_agent_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
